@@ -9,7 +9,7 @@ import * as _ from 'lodash';
 })
 export class ServiceRSYNCComponent implements OnInit {
 
-  @ViewChild('tabGroup') tabGroup;
+  @ViewChild('tabGroup', { static: true}) tabGroup;
 
   public activedTab = 'Configuration';
   public navLinks: Array < any > = [{
