@@ -7,8 +7,8 @@ import { MatProgressBar, MatButton } from '@angular/material';
   styleUrls: ['./lockscreen.component.css']
 })
 export class LockscreenComponent implements OnInit {
-  @ViewChild(MatProgressBar) progressBar: MatProgressBar;
-  @ViewChild(MatButton) submitButton: MatButton;
+  @ViewChild(MatProgressBar, { static: false}) progressBar: MatProgressBar;
+  @ViewChild(MatButton, { static: false}) submitButton: MatButton;
 
   lockscreenData = {
     password: ''
