@@ -39,7 +39,7 @@ export class EntityUtils {
                 entity.conf.isBasicMode = false;
               }
             if (!scroll) {
-              element.scrollIntoView({behavior: "instant", block: "end", inline: "nearest"});
+              element.scrollIntoView({behavior: "auto", block: "end", inline: "nearest"});
               scroll = true;
             }
           }
@@ -83,7 +83,7 @@ export class EntityUtils {
                 entity.conf.isBasicMode = false;
               }
             if (!scroll) {
-              element.scrollIntoView({behavior: "instant", block: "end", inline: "nearest"});
+              element.scrollIntoView({behavior: "auto", block: "end", inline: "nearest"});
               scroll = true;
             }
           }
